Abstracted stripes connect elements out of RenderPermissions
