+++ UserPermissions.js
@@ -40,50 +40,18 @@ class UserPermissions extends React.Component {
 
   constructor(props) {
     super(props);
-    this.state = {
-      addPermissionOpen: false,
-      searchTerm: '',
-      // handling active Permissions in state for presentation purposes only.
-    };
 
-    this.onToggleAddPermDD = this.onToggleAddPermDD.bind(this);
     this.addPermission = this.addPermission.bind(this);
     this.removePermission = this.removePermission.bind(this);
-    this.onChangeSearch = this.onChangeSearch.bind(this);
-    this.isPermAvailable = this.isPermAvailable.bind(this);
-  }
 
-  onToggleAddPermDD() {
-    const isOpen = this.state.addPermissionOpen;
-    this.setState({
-      addPermissionOpen: !isOpen,
-    });
-  }
-
-  onChangeSearch(e) {
-    const searchTerm = e.target.value;
-    this.setState({ searchTerm });
   }
 
   addPermission(perm) {
-    this.props.mutator.userPermissions.POST(perm, this.props.viewUserProps).then(() => {
-      this.onToggleAddPermDD();
-    });
+    this.props.mutator.userPermissions.POST(perm, this.props.viewUserProps);
   }
 
   removePermission(perm) {
-    console.log(this.props);
     this.props.mutator.userPermissions.DELETE(perm, this.props.viewUserProps, perm.permissionName);
-  }
-
-  isPermAvailable(perm) {
-    const permInUse = _.some(this.props.data.userPermissions, perm);
-
-    // This should be replaced with proper search when possible.
-    const nameToCompare = !perm.displayName ? perm.permissionName.toLowerCase() : perm.displayName.toLowerCase();
-    const permNotFiltered = _.includes(nameToCompare, this.state.searchTerm.toLowerCase());
-
-    return !permInUse && permNotFiltered;
   }
 
   render() {
@@ -92,14 +60,10 @@ class UserPermissions extends React.Component {
 
     return (<RenderPermissions 
               {...this.props}
-              heading="User permissions" 
-              isPermAvailable={this.isPermAvailable} 
+              heading="User Permissions" 
               addPermission={this.addPermission} 
-              onChangeSearch={this.onChangeSearch} 
               removePermission={this.removePermission} 
-              state={this.state} 
-              onToggleAddPermDD={this.onToggleAddPermDD}
-              userPermissions={userPermissions}
+              listedPermissions={userPermissions}
            ></RenderPermissions>);
       
   }
